18SJ: Move to new structure
Part of #3849
