+++ lib/engine/game/g_18_sj/meta.rb
@@ -1,0 +1,23 @@
+# frozen_string_literal: true
+
+require_relative '../meta'
+
+module Engine
+  module Game
+    module G18SJ
+      module Meta
+        include Game::Meta
+
+        DEV_STAGE = :alpha
+
+        GAME_DESIGNER = 'Örjan Wennman'
+        GAME_INFO_URL = 'https://github.com/tobymao/18xx/wiki/18SJ'
+        GAME_LOCATION = 'Sweden'
+        GAME_PUBLISHER = :self_published
+        GAME_RULES_URL = 'https://drive.google.com/file/d/1WgvqSp5HWhrnCAhAlLiTIe5oXfYtnVt9/view?usp=drivesdk'
+
+        PLAYER_RANGE = [2, 6].freeze
+      end
+    end
+  end
+end
