+++ lib/engine/game/g_18_sj/step/buy_sell_par_shares.rb
@@ -1,0 +1,17 @@
+# frozen_string_literal: true
+
+require_relative '../../../step/buy_sell_par_shares'
+
+module Engine
+  module Game
+    module G18SJ
+      module Step
+        class BuySellParShares < Engine::Step::BuySellParShares
+          def can_sell?(entity, bundle)
+            super && bundle.corporation.floated?
+          end
+        end
+      end
+    end
+  end
+end
