+++ lib/engine/game/g_18_sj/step/buy_train.rb
@@ -1,0 +1,26 @@
+# frozen_string_literal: true
+
+require_relative '../../../step/base'
+require_relative '../../../step/buy_train'
+require_relative 'buy_train_action'
+
+module Engine
+  module Game
+    module G18SJ
+      module Step
+        class BuyTrain < Engine::Step::BuyTrain
+          include BuyTrainAction
+
+          def actions(entity)
+            # If this entity has used Motala Verkstad to buy train(s) do not allow any normal train buys
+            return [] if @round.respond_to?(:premature_trains_bought) && @round.premature_trains_bought.include?(entity)
+
+            super
+          end
+
+          def do_after_buy_train_action(_action, _entity); end
+        end
+      end
+    end
+  end
+end
