+++ lib/engine/game/g_18_sj/step/buy_train_action.rb
