--- lib/engine/step/g_18_sj/buy_sell_par_shares.rb
@@ -1,15 +1,0 @@
-# frozen_string_literal: true
-
-require_relative '../buy_sell_par_shares'
-
-module Engine
-  module Step
-    module G18SJ
-      class BuySellParShares < BuySellParShares
-        def can_sell?(entity, bundle)
-          super && bundle.corporation.floated?
-        end
-      end
-    end
-  end
-end
