--- lib/engine/step/g_18_sj/route.rb
@@ -1,77 +1,0 @@
-# frozen_string_literal: true
-
-require_relative '../../game_error'
-require_relative '../route'
-
-module Engine
-  module Step
-    module G18SJ
-      class Route < Route
-        def setup
-          @game.make_sj_tokens_passable_for_electric_trains(current_entity)
-
-          super
-        end
-
-        def actions(entity)
-          return [] unless can_run_trains?(actual_entity(entity))
-
-          ACTIONS
-        end
-
-        def available_hex(entity, hex)
-          @game.graph.reachable_hexes(actual_entity(entity))[hex]
-        end
-
-        GKB_BONUS = { 3 => 50, 2 => 30, 1 => 20 }.freeze
-
-        def process_run_routes(action)
-          route = action.routes.first
-          abilities = route&.abilities
-          ability_type = abilities.first if abilities
-          used_ability = @game.abilities(action.entity, ability_type) if ability_type
-          count = used_ability&.count
-
-          super
-
-          return unless count
-
-          if count == 1
-            @game.log << "All OR bonuses for #{@game.gkb.name} used up"
-          else
-            current_amount = GKB_BONUS[count]
-            next_amount = GKB_BONUS[count - 1]
-            @game.log << "#{@game.gkb.name} bonus decrease from #{@game.format_currency(current_amount)} to "\
-              "#{@game.format_currency(next_amount)}"
-            used_ability.amount = next_amount
-          end
-        end
-
-        private
-
-        def can_run_trains?(entity)
-          entity.operator? && !entity.runnable_trains.empty? && @game.can_run_route?(entity)
-        end
-
-        def bonus_available?(entity)
-          return false unless entity.company?
-          return false if !@game.gkb ||
-            @game.gkb.closed? ||
-            entity.player != gkb_owner&.player ||
-            !current_entity.corporation? ||
-            current_entity != @gkb_owner
-
-          @game.abilities(@game.gkb, :hex_bonus, time: 'route')
-        end
-
-        def gkb_owner
-          @gkb_owner ||= @game.gkb&.owner
-        end
-
-        def actual_entity(entity)
-          bonus_available?(entity) ? entity.owner : entity
-        end
-      end
-    end
-  end
-end
